requirements.txt
pytest>=8.3.0
pytest-playwright>=0.6.0
pytest-html>=4.1.1
pytest-xdist>=3.5.0
playwright>=1.47.0
python-dotenv>=1.0.1
